use test files from resource
