+++ Tests/Reqnroll.SystemTests/Plugins/XRetryPluginTest.cs
@@ -19,50 +19,8 @@ public class XRetryPluginTest : SystemTestBase
     [TestMethod]
     public void XRetry_should_work_with_Reqnroll()
     {
-        AddScenario(
-            """
-            @retry
-            Scenario: Scenario with Retry
-                When fail for first 2 times A
-            """);
-        AddScenario(
-            """
-            @retry
-            Scenario Outline: Scenario outline with Retry
-                When fail for first 2 times <label>
-            Examples:
-                | label |
-                | B     |
-                | C     |
-            """);
-        AddBindingClass(
-            """
-            using System.Collections.Generic;
-            namespace XRetryPluginTest.StepDefinitions
-            {
-                [Binding]
-                public class XRetryPluginTestStepDefinitions
-                {
-                    private static readonly Dictionary<string, int> RetriesByLabel = new Dictionary<string, int>();
-                
-                    [When("fail for first {int} times {word}")]
-                    public void WhenFailForFirstTwoTimes(int retryCount, string label)
-                    {
-                        if (!RetriesByLabel.TryGetValue(label, out var retries))
-                        {
-                            retries = 0;
-                        }
-                        var failTest = retries < retryCount;
-                        RetriesByLabel[label] = ++retries;
-                        if (failTest)
-                        {
-                            Log.LogCustom("simulated-error", label);
-                            throw new Exception($"simulated error for {label}");
-                        }
-                    }
-                }
-            }
-            """);
+        AddFeatureFileFromResource("XRetryPlugin/XRetryPluginTestFeature.feature");
+        AddBindingClassFromResource("XRetryPlugin/XRetryPluginTestStepDefinitions.cs");
 
         ExecuteTests();
